🩹 fix(build): resolve conflict caused by Jekyll Archives plugin
The Jekyll Archives plugin was creating duplicate files in the `_site`
directory for some archive pages, resulting in unexpected contents and
possible functionality issues.

This commit fixes the issue by renaming the source file that matches
the archive path and using the front matter `redirect_to` instead of
`redirect_from`. This ensures that only one file is generated per
destination and avoids the conflict.

- Renamed `category/doc` to `category/guide`
- Used `redirect_to` in front matter of `category/guide`

Fixes: Issue #50