👷‍♂️ feat(gulp): add AMP Optimizer log
This commit adds a `console.log` statement to the Gulp task that
iterates over the HTML files of the Jekyll build and runs the AMP
Optimizer on them. The `console.log` statement prints the current
timestamp and the file name that is being optimized. This will help
to monitor the optimization process and identify any issues or errors.

See also: PR #27