⬆ chore(deps): bump semver from 5.7.1 to 5.7.2
This commit updates the semver package to version 5.7.2, which fixes a
regular expression denial of service vulnerability in the
semver.satisfies method. This vulnerability affects the
gulp-minify-inline package, which uses semver to check the version of
inline scripts and styles.

The vulnerability was reported by Snyk and assigned the identifier
GHSA-c2qf-rxjj-qqgw [1]. The fix was released in semver version 7.5.2
and backported to version 5.7.2 [2]. Updating the semver package to the
latest patch version ensures that the project is protected from this
vulnerability.

References:
[1]: https://security.snyk.io/package/npm/semver/5.7.1
[2]: https://www.npmjs.com/package/semver

Fixes for https://github.com/chriskyfung/amp-affiliately-jekyll-theme/security/dependabot/31