⬆ chore(deps): update glob-parent to fix ReDoS vulnerability
This commit resolves the glob-parent (< 5.1.2) dependency in the
gulp@4.0.2 package, which contains a high severity Regular Expression
Denial of Service (ReDoS) vulnerability in the enclosure regex. This
vulnerability could allow an attacker to cause a denial of service by
crafting a malicious glob pattern. See
https://github.com/advisories/GHSA-ww39-953v-wcq6 for more details.

Here are the changes made for the security fixes:
- Update glob-watcher for gulp:

  Overrides the glob-watcher dependency version from 5.0.5 to 6.0.0 for
  the gulp@4.02 package. This will enable the use of the latest
  glob-watcher features and improve the file watching functionality of
  gulp.

- Update vinyl-fs for gulp:

  Overrides the vinyl-fs dependency version from 3.0.3 to 4.0.0 for the
  gulp@4.0.2 package. This will allow the use of the latest vinyl-fs
  features and improve the file system operations of gulp.

See also: https://github.com/chriskyfung/amp-affiliately-jekyll-theme/security/dependabot/5