✨ feat(ui): Introduce granular cookie consent UI
This commit develops a granular cookie consent UI, composed of three key
visual components:

- a consent banner featuring "Accept All" and "Preference" buttons,
- a cookie preference modal that enables users to select their consent
  preferences, and
- a round sticky button positioned at the bottom left corner of the
  screen that allows users to reopen the consent banner.

This UI is constructed in `/consent/granular-user-consent.html` and its
corresponding CSS styles are defined in `/css/amp_consent.css` under the
`_includes` folder.

The consent UI is designed to cater to four consent purposes: Essential,
Functional, Analytics, and Advertising. This allows users to provide
consent on a per-purpose basis, enhancing user control and privacy.

The consent UI utilizes AMP (Accelerated Mobile Pages) technology,
including the use of the `amp-consent`, `amp-bind`, and `amp-script`
extensions to manage user interactions and consent states.

Here's a brief overview of what the code does:
1. Use `amp-state` to hold the consent states for different types of
   cookies (essential, functional, analytics, advertising).
2. Use `amp-script` to load the stored consent states from the local
   storage and override the default values defined in the `amp-state`.
3. Use `amp-consent` to set up a consent UI and a post consent UI. The
   consent UI includes a consent banner and a cookie preference modal.
4. The consent banner has an "Accept All" button and a "Preferences"
   button. The "Preferences" button made up with multiple instances,
   each instance is associated with a different set of consent purposes
   and is shown or hidden based on the current consent state.
5. The cookie preference modal allows users to select their consent
   preferences for different types of cookies. It also includes "Accept"
   and "Decline" buttons.
6. The post consent UI includes a button that prompts the consent UI
   when clicked.

The consent interface is incorporated into the default layout
(_layouts/default.html) using Jekyll include tags.

To ensure the CSS styles for the consent UI are loaded at the start of
the page load, we import the `amp_consent.css` into the HEAD element of
the default layout.

To make the consent UI immediately visible to users when the page loads,
we import the `granular-user-consent.html` into the default layout just
below the `<body>` tag.

see: #44