🚩 feat(amp): Make `amp-script` mandatory
This change makes the `amp-script` extension mandatory, which  is
necessary for the custom script to be able to access the localstorage,
where the stored consent states are kept.

Without this change, the custom script would not be able to read the
stored consent states and would not be able to function properly.

See: #44