✨ feat: add Privacy Policy URL configuration
Implemented a new feature that allows the user to set the URL for the
privacy policy page in the `_config.yml` file. The configuration can be
done as follows:

```yml
privacy-policy:
  url: /privacy-policy.html
```

In case the attribute isn't defined, the system will fallback to
`/privacy-policy` as the default URL.

Also, updated the link URL in the `footer-right-text.html` and
`granular-user-consent.html` templates to reflect this change. This
enhancement provides more flexibility in setting up the privacy policy
page and improves the overall user experience.

Additionally, the documentation in the `config-guide.md` has been
updated to reflect these changes and provide guidance on how to use this
new feature.