🐛 fix: update key for read amp-store from localstorage
This commit update the cookie consent UI to read the stored consent
states from the LocalStorage using the key `amp-store:{{ site.url }}`
instead of the previously used `amp-store:http://127.0.0.1:4000`.

This change provides a more accurate and reliable way of retrieving
consent states, as it is now tied to the specific site domain rather
than a local host address. This will improve the overall functionality
of the cookie consent feature and enhance the user experience.

See: #44