🔒 fix: block Disqus Comments until user accepts functional cookies
This commit adds the `data-block-on-consent-purposes="functional"`
attribute to the `amp-iframe` component used for Disqus Comments. This
change ensures that the Disqus comments feature will only function if
the user has given their consent, in compliance with GDPR and other
privacy regulations, when the consent UI is enabled. This enhances the
privacy controls of the site and respects user preferences.

The addition of this attribute does not affect the overall performance
of the site or the Disqus comments feature. It simply adds an additional
layer of user consent verification before the Disqus comments
functionality is enabled.

See: #44