🔒 fix: block addthis share buttons until user accepts functional cookies
This commit adds the `data-block-on-consent-purposes="functional"`
attribute to the `amp-addthis` component. This change ensures that the
AddThis share buttons feature will only function if the user has given
their consent, in compliance with GDPR and other privacy regulations,
when the consent UI is enabled.

This enhances the privacy controls of the site and respects user preferences.

See: #44