🔒 fix: block adsense ads until user accepts advertising cookies
This commit adds the `data-block-on-consent-purposes="advertising"`
attribute to both the `amp-ad` and `amp-auto-ads` components. This
change ensures that the Adsense ads will only be displayed if the user
has given their consent for advertising cookies, in compliance with GDPR
and other privacy regulations, when the consent UI is enabled.

See: #44