🚑 fix(amp): remove `data-ampdevmode` on non-dev environment
This commit removes the `data-ampdevmode` attribute from the `<amp-script>` component in non-development environments. This attribute is used to enable development-specific features, such as the ability to load scripts from a local server. However, it can also lead to unintended behavior in production environments, such as preventing scripts from running or causing validation errors.

By removing this attribute in non-development environments, we ensure that AMP pages are always served with the correct configuration and that scripts are executed as intended.

To address this issue, this commit adds two new options to the site configuration:
- `amp_optimizer`: Set true if you run AMP Optimizer after building the Jekyll site. Defaults to false.
- `script_hash`: A hash for the get_stored_consentStates script. Defaults to empty.

Fixes #58
See also #44