🔧 chore(editor): Config MD046 rule for markdownlint
This commit configures the MD046 rule of markdownlint extension in VS
Code settings to use the fenced style for code blocks. This style is
preferred over the indented style because it is more consistent,
readable, and compatible with other markdown parsers. This change will
help to maintain a high-quality and standard markdown documentation for
the project.