👷 build(gulp): conditionally disable sourcemaps in production
This commit modifies the `minifyCSS` Gulp task to conditionally generate sourcemaps. Sourcemaps will now only be generated if the `NODE_ENV` environment variable is not set to 'production'.

The change introduces `gulp-util` to provide a `noop()` function, which effectively bypasses the `sourcemaps.init()` and `sourcemaps.write()` pipes when `isProd` is true. This prevents unnecessary sourcemap files from being created and deployed in production builds, optimizing the build output.